% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_init.R
\docType{package}
\name{bcf-package}
\alias{bcf}
\alias{bcf-package}
\title{Compute causal effect estimates using Bayesian Adaptive Regression Trees. The methods here are those of Hahn in (citation). BART tree-based estimates under treatment and control are computed and returned for each subject. Because of an implied first "split" on the treatment variable estimates of treatment effect are allowed to be highly heterogenious and ITEs and subpopulation effects are directly computable.}
\description{
Compute causal effect estimates using Bayesian Adaptive Regression Trees. The methods here are those of Hahn in (citation). BART tree-based estimates under treatment and control are computed and returned for each subject. Because of an implied first "split" on the treatment variable estimates of treatment effect are allowed to be highly heterogenious and ITEs and subpopulation effects are directly computable.
}
\author{
\strong{Maintainer}: Christopher King \email{christopherking@wustl.edu}

}
