% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_lbart.R
\name{lbart_bcf}
\alias{lbart_bcf}
\title{Compute the BCF predictions after propensity score estimation}
\usage{
lbart_bcf(x.train, y.train, x.test = matrix(0, 0, 0), numtreated = 1L,
  extravar = TRUE, use_cauchy = FALSE, sparse = FALSE, a = 0.5, b = 1,
  augment = FALSE, rho = NULL, xinfo = matrix(0, 0, 0),
  usequants = FALSE, xinfo2 = matrix(0, 0, 0), cont = FALSE,
  rm.const = TRUE, tau.interval = 0.95, k = 2, power = 2, base = 0.95,
  binaryOffset = 0, ntree = 200L, numcut = 100L, numcut_treated = 100L,
  ntree_treated = NULL, score_index = ncol(x.train), ndpost = 1000L,
  nskip = 100L, keepevery = 1L, nkeeptrain = ndpost, nkeeptest = ndpost,
  nkeeptreedraws = ndpost, printevery = 100, transposed = FALSE)
}
\value{
The same a pbart_bcf
}
\description{
\code{lbart_bcf} returns the predicted values with treatment =0 and treatment=1 for all datapoints and test points. 
Propensity scores are treated like any other covariate. 
Unlike standard BART, two sets of trees are returned. Uses the method of Hahn_cite including Gelman's half-cauchy scaling of treatment effects and modified default prior.
}
\details{
lbart_bcf assumes that the data have been sorted so that the treated group is first. lbart is unfinished as I can't see a reason why one would use it instead of pbart
}
