% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_mc.wbart.R
\name{mc.wbart_bcf}
\alias{mc.wbart_bcf}
\title{BCF predictions after propensity score estimation with internal parallelization}
\usage{
mc.wbart_bcf(x.train, y.train, x.test = matrix(0, 0, 0), numtreated = 1L,
  extravar = TRUE, use_cauchy = FALSE, sparse = FALSE, theta = 0,
  omega = 1, a = 0.5, b = 1, augment = FALSE, rho = NULL,
  xinfo = matrix(0, 0, 0), usequants = FALSE, xinfo2 = matrix(0, 0, 0),
  cont = FALSE, rm.const = TRUE, sigest = NA, sigdf = 3,
  sigquant = 0.9, k = 2, power = 2, base = 0.95, sigmaf = NA,
  lambda = NA, fmean = mean(y.train), w = rep(1, length(y.train)),
  ntree = 200L, numcut = 100L, ntree_treated = NULL,
  score_index = ncol(x.train), ndpost = 1000L, nskip = 100L,
  keepevery = 1L, printevery = 100L, keeptrainfits = TRUE,
  transposed = FALSE, mc.cores = 2L, nice = 19L, seed = 99L)
}
\arguments{
\item{x.train}{same as wbart from BART package (except sorted based on the treatment variable)}

\item{y.train}{same as wbart from BART package (except sorted based on the treatment variable)}

\item{x.test}{same as wbart from BART package}

\item{numtreated}{Integer, how many examples are treatment = 1.}

\item{extravar}{Bool, if TRUE use heterogenious variance in the treated group}

\item{use_cauchy}{Bool, if TRUE use the cauchy over-parameterization of the variance of leaf means as in Hahn. If FALSE, standard variance prior.}

\item{sparse}{same as wbart from BART package}

\item{theta}{same as wbart from BART package}

\item{omega}{same as wbart from BART package}

\item{a}{same as wbart from BART package}

\item{b}{same as wbart from BART package}

\item{augment}{same as wbart from BART package}

\item{rho}{same as wbart from BART package}

\item{xinfo}{same as wbart from BART package}

\item{xinfo2}{same as xinfo for the treated group only}

\item{cont}{same as wbart from BART package}

\item{rm.const}{same as wbart from BART package}

\item{sigest}{same as wbart from BART package}

\item{sigdf}{same as wbart from BART package}

\item{sigquant}{same as wbart from BART package}

\item{k}{same as wbart from BART package}

\item{power}{same as wbart from BART package}

\item{base}{same as wbart from BART package}

\item{sigmaf}{same as wbart from BART package}

\item{lambda}{same as wbart from BART package}

\item{fmean}{same as wbart from BART package}

\item{w}{same as wbart from BART package}

\item{ntree}{same as wbart from BART package}

\item{numcut}{same as wbart from BART package}

\item{ntree_treated}{number of trees to use for \alpha, will set as proportional to sqrt(numtreated/n) if unspecified}

\item{score_index}{which column (row) holds the propensity score. only used to calculate approximate local average treatment effect on those with propensity score 0.5 +- 0.1}

\item{ndpost}{same as wbart from BART package}

\item{nskip}{same as wbart from BART package}

\item{keepevery}{same as wbart from BART package}

\item{printevery}{same as wbart from BART package}

\item{transposed}{same as wbart from BART package (assumes xinfo, xinfo2, numcut, numcut_treated provided)}

\item{mc.cores}{Integer, how many processes to spawn.}

\item{nice}{Integer, what nice level to apply (to prevent it locking up the system)}

\item{seed}{Integer, seed for special parallel RNG}
}
\value{
Identical to wbart_bcf
}
\description{
\code{mc.wbart_bcf} returns the (continuous) predicted values for each datapoint with treatment = 0 and treatment = 1 with BART-based prognostic score and BART-based treatment heterogenity. It just calls wbart_bcf.
}
