% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_pbart.R
\name{pbart_bcf}
\alias{pbart_bcf}
\title{Compute the BCF predictions after propensity score estimation}
\usage{
pbart_bcf(x.train, y.train, x.test = matrix(0, 0, 0), numtreated = 1L,
  extravar = TRUE, use_cauchy = FALSE, sparse = FALSE, theta = 0,
  omega = 1, a = 0.5, b = 1, augment = FALSE, rho = NULL,
  xinfo = matrix(0, 0, 0), usequants = FALSE, xinfo2 = matrix(0, 0, 0),
  cont = FALSE, rm.const = TRUE, k = 2, power = 2, base = 0.95,
  binaryOffset = 0, ntree = 50L, numcut = 100L, numcut_treated = 100L,
  ntree_treated = NULL, score_index = ncol(x.train), ndpost = 1000L,
  nskip = 100L, keepevery = 1L, nkeeptrain = ndpost, nkeeptest = ndpost,
  nkeeptreedraws = ndpost, printevery = 100L, transposed = FALSE)
}
\arguments{
\item{x.train, y.train}{same as pbart from BART package (except sorted based on the treatment variable)}

\item{x.test}{same as pbart from BART package}

\item{numtreated}{Integer, how many examples are treatment = 1.}

\item{extravar}{Not used.}

\item{use_cauchy}{Bool, if TRUE use the cauchy over-parameterization of the variance of leaf means as in Hahn. If FALSE, standard variance prior.}

\item{sparse, theta, omega, a, b, augment, rho, xinfo}{same as pbart from BART package}

\item{xinfo2}{same as xinfo for the treated group only; estimated if not provided (using cutpoints outside the support of the treated sample will cause unpredicable behavior)}

\item{cont, rm.const, k, power, base, binaryOffset, ntree, numcut}{same as pbart from BART package}

\item{numcut_treated}{same as numcut restricted to the treatment group}

\item{ntree_treated}{number of trees to use for \alpha, will set as proportional to sqrt(numtreated/n) if unspecified}

\item{score_index}{which column (row) holds the propensity score. only used to calculate approximate local average treatment effect on those with propensity score 0.5 +- 0.1}

\item{ndpost, nskip, keepevery, nkeeptrain, nkeeptest, nkeeptreedraws, printevery}{same as pbart from BART package}

\item{transposed}{same as pbart from BART package (assumes xinfo, xinfo2, numcut, numcut_treated provided)}
}
\value{
An object derived from the pbart class in package BART. It is a named list without the class attribute of BART::pbart. 
  Two sets of most return values are provided: treedraws.treated, varcount.treated, varprob.treated  contains information on \alpha(), treedraws etc information on m(). Estimated values (on the latent probit scale) with and without treatment are returned with rows representing iterations in the simulation and columns the datapoint. The estimates under treatment are stored in yhat.train.treated, yhat.test.treated. The vector of estimated residual sd (over iterations) is returned in sigma.
  Estimates of various causal parameters are returned as ate_est (\hat{y|treated} -\hat{y|untreated}), ate_sd, and ate_ci (a hpd interval) averaging over the sample, att_... averaging over the treated sample, att_modified_... averages y[treated] - \hat{y|untreated}[treated], and late_...  averages over training data where E[T|X] is within 0.5+-0.1.
}
\description{
\code{pbart_bcf} returns the (latent) predicted values for each datapoint with treatment = 0 and treatment = 1 with BART-based prognostic score and BART-based treatment heterogenity.
}
\details{
The "Bayesian Causal Forest" proposol of Hahn estimates individual treatment effects (ITE) for binary treatments with a two step procedure. First, propensity scores (E[treatment] | covariates) are estimated using T ~ BART(X) with the default prior of Chapman et al. Second, the outcomes are modeled Y ~ m(X, \hat{T}) + T*\alpha(X, \hat{T}), where m() also has a standard BART prior, and \alpha has a slightly modified prior. Propensity scores are treated like any other covariate in the estimation. This implementation is derived from the BART package. Most of the function arguments are identical to BART's similarly named functions. \code{pbart_bcf} assumes that outcomes are binary, that the data have been sorted so that the treated group is first, and that the treatment indicator is not among the covariates. A probit link function is used.
}
\seealso{
code{\link{mc.pbart_bcf}} for an internal parallel method,  code{\link{wbart_bcf}} for a method for continuous outcomes.
}
